+++ demos/petstore/src/main/java/is/codion/framework/demos/petstore/ui/PetstoreAppPanel.java
@@ -9,7 +9,6 @@ import is.codion.swing.framework.model.SwingEntityModel;
 import is.codion.swing.framework.ui.EntityApplicationPanel;
 import is.codion.swing.framework.ui.EntityPanel;
 import is.codion.swing.framework.ui.EntityPanel.PanelState;
-import is.codion.swing.framework.ui.TabbedPanelLayout;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -18,6 +17,7 @@ import java.util.Locale;
 
 import static is.codion.framework.demos.petstore.domain.Petstore.*;
 import static is.codion.swing.framework.ui.TabbedPanelLayout.detailPanelState;
+import static is.codion.swing.framework.ui.TabbedPanelLayout.splitPaneResizeWeight;
 
 public final class PetstoreAppPanel extends EntityApplicationPanel<PetstoreAppModel> {
 
@@ -39,12 +39,12 @@ public final class PetstoreAppPanel extends EntityApplicationPanel<PetstoreAppMo
 
     EntityPanel categoryPanel = new EntityPanel(categoryModel,
             new CategoryEditPanel(categoryModel.editModel()),
-            TabbedPanelLayout.splitPaneResizeWeight(0.3));
+            settings -> settings.panelLayout(splitPaneResizeWeight(0.3)));
     EntityPanel productPanel = new EntityPanel(productModel,
             new ProductEditPanel(productModel.editModel()));
     EntityPanel itemPanel = new EntityPanel(itemModel,
             new ItemEditPanel(itemModel.editModel()),
-            detailPanelState(PanelState.HIDDEN));
+            settings -> settings.panelLayout(detailPanelState(PanelState.HIDDEN)));
     EntityPanel tagItemPanel = new EntityPanel(tagItemModel,
             new TagItemEditPanel(tagItemModel.editModel()));