+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanelBuilder.java
@@ -244,8 +244,10 @@ final class EntityPanelBuilder implements EntityPanel.Builder {
   }
 
   private EntityPanel createPanel(SwingEntityModel entityModel, EntityEditPanel editPanel, EntityTablePanel tablePanel) throws Exception {
-    return panelClass().getConstructor(SwingEntityModel.class, EntityEditPanel.class, EntityTablePanel.class, PanelLayout.class)
-            .newInstance(entityModel, editPanel, tablePanel, panelLayout);
+    Consumer<EntityPanel.Settings> configure = settings -> settings.panelLayout(panelLayout);
+
+    return panelClass().getConstructor(SwingEntityModel.class, EntityEditPanel.class, EntityTablePanel.class, Consumer.class)
+            .newInstance(entityModel, editPanel, tablePanel, configure);
   }
 
   private EntityEditPanel createEditPanel(SwingEntityEditModel editModel) {